chore: move modules under `nock` to the root
