+++ src/hash-generator/index.ts
