+++ src/hash-generator/utils.ts
@@ -1,4 +1,4 @@
-import { IGenerator } from '../../interfaces';
+import { IGenerator } from '../interfaces';
 import {
   InputGenerateHash,
   JestGenerateHash,
